TST/MAINT: Refactor tmp_file fixture to fixtures module with variable content
